Additional tests for check_syntax
- Stop usage of C-style string data and functions; can be viewed as an exhaustive extension of #2911.
- Ensure that "#pragma one" is used instead of the historical "#ifndef __headerpath__", as introduced in #2871.
- Enforce use of nested namespaces, as first introduced in #2652.
- Enforce explicit resolution of MR::abs() for templated data types; where data types are exclusively scalar floating-point, std::fabs() is instead used.
- Preclude use of C-style casts, as proposed in #3010.
- Add to MR namespace "NaNF" and "InfF" as floating-point variants of the already-defined default_type "NaN" and "Inf", and preclude use of macros "NAN" and "INFINITY" in favour of either one of those or std::numeric_limits<>, as proposed in #3192.
- Preclude use of "NULL" macro, replacing with nullptr.

+++ cpp/core/dwi/tractography/algorithms/iFOD1.h
@@ -79,8 +79,8 @@ public:
     }
 
     ~Shared() {
-      mean_samples /= double(num_proc);
-      mean_truncations /= double(num_proc);
+      mean_samples /= static_cast<double>(num_proc);
+      mean_truncations /= static_cast<double>(num_proc);
       INFO("mean number of samples per step = " + str(mean_samples));
       if (mean_truncations) {
         INFO("mean number of steps between rejection sampling truncations = " + str(1.0 / mean_truncations));
@@ -119,8 +119,9 @@ public:
   }
 
   ~iFOD1() {
-    S.update_stats(calibrate_list.size() + float(mean_sample_num) / float(num_sample_runs),
-                   float(num_truncations) / float(num_sample_runs),
+    S.update_stats(calibrate_list.size() +
+                       (static_cast<double>(mean_sample_num) / static_cast<double>(num_sample_runs)),
+                   static_cast<double>(num_truncations) / static_cast<double>(num_sample_runs),
                    max_truncation);
   }
